+++ lib/processor/process_map.go
@@ -21,6 +21,7 @@
 package processor
 
 import (
+	"encoding/json"
 	"fmt"
 
 	"github.com/Jeffail/benthos/lib/log"
@@ -35,7 +36,9 @@ import (
 
 func init() {
 	Constructors[TypeProcessMap] = TypeSpec{
-		constructor: NewProcessMap,
+		constructor: func(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
+			return NewProcessMap(conf.ProcessMap, mgr, log, stats)
+		},
 		description: `
 A processor that extracts and maps fields from the original payload into new
 objects, applies a list of processors to the newly constructed objects, and
@@ -99,28 +102,7 @@ processing they will be correctly aligned with the original batch. However, the
 ordering of premapped message parts as they are sent through processors are not
 guaranteed to match the ordering of the original batch.`,
 		sanitiseConfigFunc: func(conf Config) (interface{}, error) {
-			var err error
-			condConfs := make([]interface{}, len(conf.ProcessMap.Conditions))
-			for i, cConf := range conf.ProcessMap.Conditions {
-				if condConfs[i], err = condition.SanitiseConfig(cConf); err != nil {
-					return nil, err
-				}
-			}
-			procConfs := make([]interface{}, len(conf.ProcessMap.Processors))
-			for i, pConf := range conf.ProcessMap.Processors {
-				if procConfs[i], err = SanitiseConfig(pConf); err != nil {
-					return nil, err
-				}
-			}
-			return map[string]interface{}{
-				"parts":            conf.ProcessMap.Parts,
-				"conditions":       condConfs,
-				"premap":           conf.ProcessMap.Premap,
-				"premap_optional":  conf.ProcessMap.PremapOptional,
-				"postmap":          conf.ProcessMap.Postmap,
-				"postmap_optional": conf.ProcessMap.PostmapOptional,
-				"processors":       procConfs,
-			}, nil
+			return conf.ProcessMap.Sanitise()
 		},
 	}
 }
@@ -152,6 +134,63 @@ func NewProcessMapConfig() ProcessMapConfig {
 	}
 }
 
+// Sanitise the configuration into a minimal structure that can be printed
+// without changing the intent.
+func (p ProcessMapConfig) Sanitise() (map[string]interface{}, error) {
+	var err error
+	condConfs := make([]interface{}, len(p.Conditions))
+	for i, cConf := range p.Conditions {
+		if condConfs[i], err = condition.SanitiseConfig(cConf); err != nil {
+			return nil, err
+		}
+	}
+	procConfs := make([]interface{}, len(p.Processors))
+	for i, pConf := range p.Processors {
+		if procConfs[i], err = SanitiseConfig(pConf); err != nil {
+			return nil, err
+		}
+	}
+	return map[string]interface{}{
+		"parts":            p.Parts,
+		"conditions":       condConfs,
+		"premap":           p.Premap,
+		"premap_optional":  p.PremapOptional,
+		"postmap":          p.Postmap,
+		"postmap_optional": p.PostmapOptional,
+		"processors":       procConfs,
+	}, nil
+}
+
+//------------------------------------------------------------------------------
+
+// UnmarshalJSON ensures that when parsing configs that are in a slice the
+// default values are still applied.
+func (p *ProcessMapConfig) UnmarshalJSON(bytes []byte) error {
+	type confAlias ProcessMapConfig
+	aliased := confAlias(NewProcessMapConfig())
+
+	if err := json.Unmarshal(bytes, &aliased); err != nil {
+		return err
+	}
+
+	*p = ProcessMapConfig(aliased)
+	return nil
+}
+
+// UnmarshalYAML ensures that when parsing configs that are in a slice the
+// default values are still applied.
+func (p *ProcessMapConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	type confAlias ProcessMapConfig
+	aliased := confAlias(NewProcessMapConfig())
+
+	if err := unmarshal(&aliased); err != nil {
+		return err
+	}
+
+	*p = ProcessMapConfig(aliased)
+	return nil
+}
+
 //------------------------------------------------------------------------------
 
 // ProcessMap is a processor that applies a list of child processors to a new
@@ -179,13 +218,13 @@ type ProcessMap struct {
 
 // NewProcessMap returns a ProcessField processor.
 func NewProcessMap(
-	conf Config, mgr types.Manager, log log.Modular, stats metrics.Type,
-) (Type, error) {
+	conf ProcessMapConfig, mgr types.Manager, log log.Modular, stats metrics.Type,
+) (*ProcessMap, error) {
 	nsStats := metrics.Namespaced(stats, "processor.process_map")
 	nsLog := log.NewModule(".processor.process_map")
 
 	var children []Type
-	for _, pconf := range conf.ProcessMap.Processors {
+	for _, pconf := range conf.Processors {
 		proc, err := New(pconf, mgr, nsLog, nsStats)
 		if err != nil {
 			return nil, err
@@ -194,7 +233,7 @@ func NewProcessMap(
 	}
 
 	var conditions []types.Condition
-	for _, cconf := range conf.ProcessMap.Conditions {
+	for _, cconf := range conf.Conditions {
 		cond, err := condition.New(cconf, mgr, nsLog, nsStats)
 		if err != nil {
 			return nil, err
@@ -203,7 +242,7 @@ func NewProcessMap(
 	}
 
 	p := &ProcessMap{
-		parts: conf.ProcessMap.Parts,
+		parts: conf.Parts,
 
 		children: children,
 
@@ -226,10 +265,10 @@ func NewProcessMap(
 		mapper.OptSetLogger(nsLog),
 		mapper.OptSetStats(nsStats),
 		mapper.OptSetConditions(conditions),
-		mapper.OptSetReqMap(conf.ProcessMap.Premap),
-		mapper.OptSetOptReqMap(conf.ProcessMap.PremapOptional),
-		mapper.OptSetResMap(conf.ProcessMap.Postmap),
-		mapper.OptSetOptResMap(conf.ProcessMap.PostmapOptional),
+		mapper.OptSetReqMap(conf.Premap),
+		mapper.OptSetOptReqMap(conf.PremapOptional),
+		mapper.OptSetResMap(conf.Postmap),
+		mapper.OptSetOptResMap(conf.PostmapOptional),
 	); err != nil {
 		return nil, err
 	}
@@ -242,6 +281,39 @@ func NewProcessMap(
 // ProcessMessage applies the processor to a message, either creating >0
 // resulting messages or a response to be sent back to the message source.
 func (p *ProcessMap) ProcessMessage(msg types.Message) ([]types.Message, types.Response) {
+	alignedResult, err := p.CreateResult(msg)
+	if err != nil {
+		msgs := [1]types.Message{msg}
+		return msgs[:], nil
+	}
+
+	result := msg.Copy()
+	if err = p.OverlayResult(result, alignedResult); err != nil {
+		msgs := [1]types.Message{msg}
+		return msgs[:], nil
+	}
+
+	msgs := [1]types.Message{result}
+	return msgs[:], nil
+}
+
+// TargetsUsed returns a list of target dependencies of this processor derived
+// from its premap and premap_optional fields.
+func (p *ProcessMap) TargetsUsed() []string {
+	return p.mapper.TargetsUsed()
+}
+
+// TargetsProvided returns a list of targets provided by this processor derived
+// from its postmap and postmap_optional fields.
+func (p *ProcessMap) TargetsProvided() []string {
+	return p.mapper.TargetsProvided()
+}
+
+// CreateResult performs reduction and child processors to a payload and returns
+// the result. The resulting message should have the same dimension as the
+// payload, where reduced indexes are nil. This result can be overlayed onto the
+// original message in order to complete the map.
+func (p *ProcessMap) CreateResult(msg types.Message) (types.Message, error) {
 	p.mCount.Incr(1)
 	p.mCountParts.Incr(int64(msg.Len()))
 
@@ -258,15 +330,15 @@ func (p *ProcessMap) ProcessMessage(msg types.Message) ([]types.Message, types.R
 		p.mErr.Incr(1)
 		p.mErrPre.Incr(1)
 		p.log.Errorf("Failed to map request: %v\n", err)
-		msgs := [1]types.Message{msg}
-		return msgs[:], nil
+		return nil, err
 	}
 
 	if mappedMsg.Len() == 0 {
 		p.mSkipped.Incr(1)
 		p.mSkippedParts.Incr(int64(msg.Len()))
-		msgs := [1]types.Message{msg}
-		return msgs[:], nil
+		parts := make([]types.Part, msg.Len())
+		mapMsg.SetAll(parts)
+		return mapMsg, nil
 	}
 
 	var procResults []types.Message
@@ -274,8 +346,7 @@ func (p *ProcessMap) ProcessMessage(msg types.Message) ([]types.Message, types.R
 		p.mErrProc.Incr(1)
 		p.mErr.Incr(1)
 		p.log.Errorf("Processors failed: %v\n", err)
-		msgs := [1]types.Message{msg}
-		return msgs[:], nil
+		return nil, err
 	}
 
 	i := 0
@@ -292,24 +363,25 @@ func (p *ProcessMap) ProcessMessage(msg types.Message) ([]types.Message, types.R
 		p.mErrPost.Incr(1)
 		p.mErr.Incr(1)
 		p.log.Errorf("Postmap failed: %v\n", err)
-		msgs := [1]types.Message{msg}
-		return msgs[:], nil
+		return nil, err
 	}
 
-	result := msg.Copy()
-	if err = p.mapper.MapResponses(result, alignedResult); err != nil {
+	return alignedResult, nil
+}
+
+// OverlayResult attempts to merge the result of a process_map with the original
+//  payload as per the map specified in the postmap and postmap_optional fields.
+func (p *ProcessMap) OverlayResult(payload, response types.Message) error {
+	if err := p.mapper.MapResponses(payload, response); err != nil {
 		p.mErrPost.Incr(1)
 		p.mErr.Incr(1)
 		p.log.Errorf("Postmap failed: %v\n", err)
-		msgs := [1]types.Message{msg}
-		return msgs[:], nil
+		return err
 	}
 
 	p.mSent.Incr(1)
-	p.mSentParts.Incr(int64(result.Len()))
-
-	msgs := [1]types.Message{result}
-	return msgs[:], nil
+	p.mSentParts.Incr(int64(payload.Len()))
+	return nil
 }
 
 func processMap(mappedMsg types.Message, processors []Type) ([]types.Message, error) {